combined ecliptic and galactic stackers
