+++ python/lsst/sims/maf/stackers/coordStackers.py
@@ -1,11 +1,28 @@
 import numpy as np
 import ephem
+from lsst.sims.utils import _galacticFromEquatorial
+
 from .baseStacker import BaseStacker
 from .ditherStackers import wrapRA
 
+__all__ = ['GalacticStacker',
+           'EclipticStacker', 'mjd2djd']
 
-__all__ = ['EclipticStacker', 'mjd2djd']
+class GalacticStacker(BaseStacker):
+    """
+    Stack on the galactic coordinates of each pointing.
+    """
+    def __init__(self, mjdCol='expMJD', raCol='fieldRA',decCol='fieldDec'):
 
+        self.colsReq = [raCol,decCol]
+        self.colsAdded = ['gall','galb']
+        self.units = ['radians', 'radians']
+        self.raCol = raCol
+        self.decCol=decCol
+
+    def _run(self,simData):
+        simData['gall'], simData['galb'] = _galacticFromEquatorial(simData[self.raCol], simData[self.decCol])
+        return simData
 
 def mjd2djd(mjd):
     """
@@ -31,9 +48,7 @@ class EclipticStacker(BaseStacker):
         self.raCol = raCol
         self.decCol=decCol
 
-    def run(self,simData):
-        simData=self._addStackers(simData)
-
+    def _run(self, simData):
         for i in np.arange(simData.size):
             coord = ephem.Equatorial(simData[self.raCol][i],simData[self.decCol][i], epoch=2000)
             ecl = ephem.Ecliptic(coord)
@@ -46,5 +61,4 @@ class EclipticStacker(BaseStacker):
                 simData['eclipLon'][i] = lon
             else:
                 simData['eclipLon'][i] = ecl.lon
-
         return simData
